update workflows in docs tab and the linking in reference and examples
