+++ docs/src/content/en/docs/workflows-legacy/_meta.tsx
@@ -1,0 +1,14 @@
+import { Tag } from "@/components/tag";
+const meta = {
+  overview: "Overview",
+  steps: "Steps",
+  "control-flow": "Control Flow",
+  variables: "Variables",
+  "suspend-and-resume": "Suspend & Resume",
+  "dynamic-workflows": "Dynamic Workflows",
+  "error-handling": "Error Handling",
+  "nested-workflows": "Nested Workflows",
+  "runtime-variables": "Runtime/Dynamic Variables",
+};
+
+export default meta;
