--- docs/src/content/en/docs/workflows-vnext/_meta.tsx
@@ -1,11 +1,0 @@
-import { Tag } from "@/components/tag";
-const meta = {
-  overview: <Tag text="new">Overview</Tag>,
-  "flow-control": "Flow Control",
-  "suspend-and-resume": "Suspend & Resume",
-  "input-data-mapping": "Input Data Mapping",
-  "using-with-agents-and-tools": "Using with Agents and Tools",
-  "inngest-workflow": <Tag text="experimental">Inngest Workflow</Tag>,
-};
-
-export default meta;
