+++ docs/src/content/en/docs/workflows/inngest-workflow.mdx
@@ -1,6 +1,6 @@
 ---
-title: "Inngest Workflows | Workflows (vNext) | Mastra Docs"
-description: "Inngest workflow allows you to run Mastra vNext workflows with Inngest"
+title: "Inngest Workflows | Workflows | Mastra Docs"
+description: "Inngest workflow allows you to run Mastra workflows with Inngest"
 ---
 
 # Inngest Workflow
@@ -169,7 +169,7 @@ const inngest = new Inngest({
 
 // Configure Mastra with the workflow and Inngest API endpoint
 export const mastra = new Mastra({
-  vnext_workflows: {
+  workflows: {
     incrementWorkflow
   },
   server: {
@@ -198,7 +198,7 @@ const srv = serve({
 })
 
 // Get the workflow, create a run, and start it with an initial value
-const workflow = mastra.vnext_getWorkflow('incrementWorkflow')
+const workflow = mastra.getWorkflow('incrementWorkflow')
 const run = workflow.createRun({})
 const result = await run.start({ inputData: { value: 5 } })
 console.dir(result, { depth: null })
