+++ docs/src/content/en/examples/workflows/workflow-variables.mdx
@@ -204,4 +204,4 @@ This example demonstrates several important concepts about workflow variables:
 
 5. **Explicit Data Dependencies**: By defining input schemas and using variable mappings, the data dependencies between steps are made explicit and clear.
 
-For more information on workflow variables, see the [Workflow Variables documentation](../../docs/workflows/variables.mdx).
+For more information on workflow variables, see the [Workflow Variables documentation](../../docs/workflows-legacy/variables.mdx).
