+++ docs/src/content/en/reference/workflows/afterEvent.mdx
@@ -73,5 +73,5 @@ workflow
 
 - [Event-Driven Workflows](./events.mdx)
 - [resumeWithEvent()](./resumeWithEvent.mdx)
-- [Suspend and Resume](../../docs/workflows/suspend-and-resume.mdx)
+- [Suspend and Resume](../../docs/workflows-legacy/suspend-and-resume.mdx)
 - [Workflow Class](./workflow.mdx)
