+++ docs/src/content/en/reference/workflows/commit.mdx
@@ -30,4 +30,4 @@ workflow.step(stepA).then(stepB).commit();
 - [Branching Paths example](../../examples/workflows/branching-paths.mdx)
 - [Workflow Class Reference](./workflow.mdx)
 - [Step Reference](./step-class.mdx)
-- [Control Flow Guide](../../docs/workflows/control-flow.mdx)
+- [Control Flow Guide](../../docs/workflows-legacy/control-flow.mdx)
