+++ docs/src/content/en/reference/workflows/else.mdx
@@ -64,5 +64,5 @@ try {
 
 - [if Reference](./if.mdx)
 - [then Reference](./then.mdx)
-- [Control Flow Guide](../../docs/workflows/control-flow.mdx)
+- [Control Flow Guide](../../docs/workflows-legacy/control-flow.mdx)
 - [Step Condition Reference](./step-condition.mdx)
