+++ docs/src/content/en/reference/workflows/events.mdx
@@ -297,7 +297,7 @@ runWorkflow().catch(console.error);
 
 ## Related
 
-- [Suspend and Resume in Workflows](../../docs/workflows/suspend-and-resume.mdx)
+- [Suspend and Resume in Workflows](../../docs/workflows-legacy/suspend-and-resume.mdx)
 - [Workflow Class Reference](./workflow.mdx)
 - [Resume Method Reference](./resume.mdx)
 - [Watch Method Reference](./watch.mdx)
