+++ docs/src/content/en/reference/workflows/if.mdx
@@ -104,5 +104,5 @@ try {
 
 - [else Reference](./else.mdx)
 - [then Reference](./then.mdx)
-- [Control Flow Guide](../../docs/workflows/control-flow.mdx)
+- [Control Flow Guide](../../docs/workflows-legacy/control-flow.mdx)
 - [Step Condition Reference](./step-condition.mdx)
