+++ docs/src/content/en/reference/workflows/resume.mdx
@@ -152,7 +152,7 @@ try {
 
 ## Related
 
-- [Suspend and Resume](../../docs/workflows/suspend-and-resume.mdx)
+- [Suspend and Resume](../../docs/workflows-legacy/suspend-and-resume.mdx)
 - [`suspend` Reference](./suspend.mdx)
 - [`watch` Reference](./watch.mdx)
 - [Workflow Class Reference](./workflow.mdx)
