+++ docs/src/content/en/reference/workflows/resumeWithEvent.mdx
@@ -128,6 +128,6 @@ await start({ triggerData: { requestId: "auto-123" } });
 
 - [Event-Driven Workflows](./events.mdx)
 - [afterEvent()](./afterEvent.mdx)
-- [Suspend and Resume](../../docs/workflows/suspend-and-resume.mdx)
+- [Suspend and Resume](../../docs/workflows-legacy/suspend-and-resume.mdx)
 - [resume()](./resume.mdx)
 - [watch()](./watch.mdx)
