+++ docs/src/content/en/reference/workflows/snapshots.mdx
@@ -204,4 +204,4 @@ run.watch(async ({ activePaths }) => {
 - [Suspend Function Reference](./suspend.mdx)
 - [Resume Function Reference](./resume.mdx)
 - [Watch Function Reference](./watch.mdx)
-- [Suspend and Resume Guide](../../docs/workflows/suspend-and-resume.mdx)
+- [Suspend and Resume Guide](../../docs/workflows-legacy/suspend-and-resume.mdx)
