+++ docs/src/content/en/reference/workflows/step-class.mdx
@@ -96,5 +96,5 @@ const processOrder = new Step({
 ## Related
 
 - [Workflow Reference](./workflow.mdx)
-- [Step Configuration Guide](../../docs/workflows/steps.mdx)
-- [Control Flow Guide](../../docs/workflows/control-flow.mdx)
+- [Step Configuration Guide](../../docs/workflows-legacy/steps.mdx)
+- [Control Flow Guide](../../docs/workflows-legacy/control-flow.mdx)
