+++ docs/src/content/en/reference/workflows/step-condition.mdx
@@ -134,4 +134,4 @@ This query syntax allows for readable conditional logic for determining whether 
 
 - [Step Options Reference](./step-options.mdx)
 - [Step Function Reference](./step-function.mdx)
-- [Control Flow Guide](../../docs/workflows/control-flow.mdx)
+- [Control Flow Guide](../../docs/workflows-legacy/control-flow.mdx)
