+++ docs/src/content/en/reference/workflows/step-function.mdx
@@ -87,7 +87,7 @@ workflow.step({
 
 ## Related
 
-- [Basic Usage with Step Instance](../../docs/workflows/steps.mdx)
+- [Basic Usage with Step Instance](../../docs/workflows-legacy/steps.mdx)
 - [Step Class Reference](./step-class.mdx)
 - [Workflow Class Reference](./workflow.mdx)
-- [Control Flow Guide](../../docs/workflows/control-flow.mdx)
+- [Control Flow Guide](../../docs/workflows-legacy/control-flow.mdx)
