+++ docs/src/content/en/reference/workflows/step-options.mdx
@@ -62,8 +62,8 @@ workflow.step(processOrder, {
 
 ## Related
 
-- [Path Comparison](../../docs/workflows/control-flow.mdx)
+- [Path Comparison](../../docs/workflows-legacy/control-flow.mdx)
 - [Step Function Reference](./step-function.mdx)
 - [Step Class Reference](./step-class.mdx)
 - [Workflow Class Reference](./workflow.mdx)
-- [Control Flow Guide](../../docs/workflows/control-flow.mdx)
+- [Control Flow Guide](../../docs/workflows-legacy/control-flow.mdx)
