+++ docs/src/content/en/reference/workflows/step-retries.mdx
@@ -193,4 +193,4 @@ You can monitor retry attempts in your logs. Mastra logs retry-related events at
 
 - [Step Class Reference](./step-class.mdx)
 - [Workflow Configuration](./workflow.mdx)
-- [Error Handling in Workflows](../../docs/workflows/error-handling.mdx)
+- [Error Handling in Workflows](../../docs/workflows-legacy/error-handling.mdx)
