+++ docs/src/content/en/reference/workflows/suspend.mdx
@@ -65,6 +65,6 @@ const reviewStep = new Step({
 
 ### Related
 
-- [Suspend & Resume Workflows](../../docs/workflows/suspend-and-resume.mdx)
+- [Suspend & Resume Workflows](../../docs/workflows-legacy/suspend-and-resume.mdx)
 - [.resume()](./resume.mdx)
 - [.watch()](./watch.mdx)
