+++ docs/src/content/en/reference/workflows/then.mdx
@@ -69,4 +69,4 @@ try {
 - [step Reference](./step-class.mdx)
 - [after Reference](./after.mdx)
 - [Sequential Steps Example](../../examples/workflows/sequential-steps.mdx)
-- [Control Flow Guide](../../docs/workflows/control-flow.mdx)
+- [Control Flow Guide](../../docs/workflows-legacy/control-flow.mdx)
