+++ docs/src/content/en/reference/workflows/until.mdx
@@ -164,5 +164,5 @@ const result = await run.start({
 ## Related
 
 - [.while()](./while.mdx) - Loop while a condition is true
-- [Control Flow Guide](../../docs/workflows/control-flow.mdx)
+- [Control Flow Guide](../../docs/workflows-legacy/control-flow.mdx)
 - [Workflow Class Reference](./workflow.mdx)
