+++ docs/src/content/en/reference/workflows/while.mdx
@@ -164,5 +164,5 @@ const result = await run.start({
 ## Related
 
 - [.until()](./until.mdx) - Loop until a condition becomes true
-- [Control Flow Guide](../../docs/workflows/control-flow.mdx)
+- [Control Flow Guide](../../docs/workflows-legacy/control-flow.mdx)
 - [Workflow Class Reference](./workflow.mdx)
